STRWEB-41 omit any mention of react-githubish-mentions
This is really just a clean-up job; in fact
`@folio/react-githubish-mentions` was not even part of the initial
commit for `package.json`, indicating it must have been removed from
`stripes-core` before this repo was split off. Anyway: we had a
lingering test case that mentioned it, but it was not a great test to
begin with, amounting to, "make sure ${items} are present in ${list}",
which just meant we had `${items}` hard-coded in two places instead of
one.

Refs STRWEB-41