STRWEB-102 forward stripes-config values to service-worker
Use a virtual module to forward some stripes-config values to the
service worker when it is bundled. These values must be compiled in
because global variables in a service worker are volatile and are thus
reset to their default (null) values on each sleep/wake cycle. Thus, the
only way to make these values permanently available is to compile them
in.

Refs STRWEB-102, STCOR-759