STRWEB-110 ignore order of CSS imports
Add `@folio/consortia-settings` to a platform build generates tens of
thousands of lines of warnings form mini-css-extract-plugin about its
inability to determine what order to assemble its CSS rules in because
different files have the imports in a different order.

It seems likely we don't have to worry about this since per-module CSS
will be scoped to that module by postcss, and shared CSS will just be in
stripes-components, but this could be riskier than I realize, and a
better approach may be to try to figure out what is happening in
`@folio/consortia-settings` that is different than every other module
we've added to the platform so far.

Refs STRWEB-110