STRWEB-113 expose 'build' command publicly
**DRAFT** **DRAFT** **DRAFT**

Expose `build` as a `bin` script, allowing a platform to directly depend
on this module and call the `build` API to generate a bundle without
pulling in all the stripes-cli deps that are unrelated to producing a
production bundle.

CSS isn't being correctly bundled/handled here, but the bundle is
otherwise functional. Hopefully, this is just a config glitch.

Refs STRWEB-113