STRWEB-124 extract sourcemaps from pre-transpiled packages
Extract sourcemaps from pre-transpiled packages. This was unnecessary
previously because stripes-webpack was responsible for both
transpilation and bundling of all `@folio`-scoped packages in a
platform. Now, some packages are transpiled on publish, so we need this
change to consume their sourcemaps.

We may want to set include/exclude rules to only include sourcemaps for
packages scoped to `@folio`.

Refs STRWEB-124