Revert "STRWEB-127 bump NodeJS to v20 (#155)"
This reverts commit 095f8eff4b3a44ed202c64bf70a4d9b5eab5be84.

The Jenkins nodes where we build platform-complete don't yet have NodeJS
v20 and we don't know when it will be available. In the interest of
making other changes (which, AFAIK, do directly not rely on NodeJS v20)
available ASAP, we are reverting this change until our build nodes can
support it.

Refs STRWEB-127