STRWEB-134 prune `stripes.js` and its deps (#166)
`./stripes.js` appears to be a legacy import from stripes-core and
hasn't been touched since. I think stripes-core's tools were completely
overhauled in their conversion to becoming stripes-cli, and this script
just got lost along the way.

`commander` is worth considering as a well-maintained alternative to
`yargs` in stripes-cli, but that's a whole different project.

Refs STRWEB-134