STRWEB-137 optionally support lazy loading (#168)
Optionally support lazy loading when constructing a bundle.

When constructing a bundle with `--lazy`, implement `getModule` with
`React.lazy` for all stripes modules except handlers, which define
static methods that are accessed outside the React render lifecycle, and
which are lost when the module is loaded via `lazy(() => import(...))`
(STRWEB-52).

Refs STRWEB-137