STRWEB-137 provide getDynamicModule()
This is a slight pivot from the original intent of STRWEB-137 to revisit
lazy-loading. Instead, it provides `getDynamicModule()`, an analog of
the legacy `getModule()` call that returns a module synchronously via
`require()` that return a module asynchronously via `import()`.

This allows consumers to decide whether to implement `lazy()` (which
will extract the default export automatically) or to load and extract it
separately. The latter is important in the context of event-handler and
plugin modules that must be loaded without suspense because they are
accessed statically during render.

Refs STRWEB-137